packages/api/pyproject.toml
[tool.poetry]
name = "loocbooc-api"
version = "0.1.0"
description = "Loocbooc Platform API — Universal Garment Identifier system and virtual try-on backend"
authors = ["Loocbooc Engineering <eng@loocbooc.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
alembic = "^1.13.0"
asyncpg = "^0.29.0"
redis = {extras = ["hiredis"], version = "^5.0.0"}
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
anthropic = "^0.25.0"
boto3 = "^1.34.0"
structlog = "^24.1.0"
httpx = "^0.27.0"
Pillow = "^10.2.0"
python-magic = "^0.4.27"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"
factory-boy = "^3.3.0"
faker = "^24.0.0"
black = "^24.2.0"
ruff = "^0.3.0"
mypy = "^1.9.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
